LLM/Q2/requirements.txt
requests
beautifulsoup4
Pillow
openai
python-dotenv